Correctly import ES6 modules in Node
TypeScript provides no means of building files with the .mjs extension.
Recent versions of NodeJS will import .js files as modules, but only if
the package.json file includes "type": "module". Refer to NodeJS
documentation for more details:

https://nodejs.org/api/esm.html#esm_enabling

Initial search for a way to do this in TypeScript ended with the comment
on TypeScript's GitHub issue tracker linked below. This made it clear
that relying on "experimental" features in node would be the only way to
properly load ES6 modules.

https://github.com/microsoft/TypeScript/issues/18442#issuecomment-581738714