Define Angle class for degree/radian conversions
Unfortunately TypeScript doesn't make it easy to use objects with
primitive operators (like +) -- turns out type coercion is a sticky
subject. Using [Symbol.toPrimitive](), .valueOf(), and similar is
possible with type assertion but seems hackey and is more verbose
than just explicitly calling .valueOf() or similar.

This class may be superfluous and possibly detrimental to performance,
but it is useful for development & debugging at least.