Generate cylinder vertex & element data for WebGL
Vertices are generated for a cylinder of given diremsions and "angular
resolution". The procedure is: 1) generate distinct vertices, 2) make
unique copies of vertices shared by non-coplanar elements, and 3) record
indices of vertices grouped by element.

Although the process is logically clear and has served as a valuable
exercise, the implementation occurs in many intermediate steps that
could probably be optimized further.