Test simulation loop over multiple runs
Calculations appear to be working as expected. Also note that the
position of the earth seems accurate - mean anomaly, printed in radians,
is roughly 93 degrees, which seems reasonable given that it's late march
and reference epoch is defined at the beginning of january.

Throttling of setTimeout() callbacks is evident in dt values.